proactive-project/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.30.0

# Qwen3-VL
qwen-vl-utils>=0.0.8

# OpenAI API (GPT-4o)
openai>=1.30.0

# Video processing
opencv-python>=4.8.0
decord>=0.6.0
Pillow>=10.0.0

# Evaluation
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Memory (embedding + retrieval)
faiss-cpu>=1.7.4

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0
